Remove unnecessary parses to float

+++ Assignment 2/MatrixTransformations/Form1.cs
@@ -169,15 +169,15 @@ namespace MatrixTransformations
             vb = cube.vertexbuffer;
 
             // Scale the cube.
-            Matrix scaled = Matrix.ScaleMatrix((float)variables.scale);
+            Matrix scaled = Matrix.ScaleMatrix(variables.scale);
 
             // Rotate the cube.
-            Matrix rotateX = Matrix.RotateMatrixX((float)variables.rotateX);
-            Matrix rotateY = Matrix.RotateMatrixX((float)variables.rotateY);
-            Matrix rotateZ = Matrix.RotateMatrixX((float)variables.rotateZ);
+            Matrix rotateX = Matrix.RotateMatrixX(variables.rotateX);
+            Matrix rotateY = Matrix.RotateMatrixX(variables.rotateY);
+            Matrix rotateZ = Matrix.RotateMatrixX(variables.rotateZ);
 
             // Translate the cube, should always be last!
-            Matrix translate = Matrix.TranslateMatrix(new Vector((float)variables.translateX, (float)variables.translateY, (float)variables.translateZ, 0));
+            Matrix translate = Matrix.TranslateMatrix(new Vector(variables.translateX, variables.translateY, variables.translateZ, 0));
 
             // Multiple them all up.
             Matrix total = scaled * rotateX * rotateY * rotateZ * translate;
@@ -304,29 +304,29 @@ namespace MatrixTransformations
 
                 // If arrows left/right are pressed, change translateX.
                 case Keys.Right:
-                    variables.translateX += 0.1;
+                    variables.translateX += 0.1f;
                     break;
 
                 case Keys.Left:
-                    variables.translateX -= 0.1;
+                    variables.translateX -= 0.1f;
                     break;
 
                 // If arrows down/up are pressed, change translateY.
                 case Keys.Up:
-                    variables.translateY += 0.1;
+                    variables.translateY += 0.1f;
                     break;
 
                 case Keys.Down:
-                    variables.translateY -= 0.1;
+                    variables.translateY -= 0.1f;
                     break;
 
                 // If arrows PageDown/PageUp are pressed, change translateZ.
                 case Keys.PageUp:
-                    variables.translateZ += 0.1;
+                    variables.translateZ += 0.1f;
                     break;
 
                 case Keys.PageDown:
-                    variables.translateZ -= 0.1;
+                    variables.translateZ -= 0.1f;
                     break;
 
                 // If X/x is pressed, change RotateX.
